Use float64 instead of float32 for storing takum8 and takum16 values
While it still makes sense to store takum8 and takum16 _logarithmic_
values (i.e. the exponents) in float32, once we want to calculate
the real represented value the dynamic range of float32 is insufficient
to cover the entire range of possible takum values.

Thus all use of single precision functions is incorrect, and we adapt
the code to drop them all in favour of double precision functions where
necessary.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>