Don't redundantly compute logarithms during addition and subtraction
We use recursion to reduce all possible input sign cases to an addition
or subtraction of positive numbers. Leading up to it, we redundantly
compute the logarithmic values in each recursion step, even though we
need these results only at the lowest level. In total the recursion is
at most 1 deep, which amounted to one redundant computation.

As addition and subtraction are standard operations, this makes a huge
difference though.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>