takum8: Properly round within the characteristic
An oversight in codec_takum8_from_s_and_l() clamps takum8 down even
when we actually want to round up in case the characteristic extends
beyond the 8 available bits.

We could solve this by adding more conditions, more smart shifts and
more complexity, but I think it should be faster to simply generate
a takum16 and round it to a takum8. It also means less code to worry
about.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>