Properly implement saturation arithmetic for math functions
By convention we never under- or overflow and only really accept
infinities for 'true' infinities that we subsequently cast to NaN.

We approach this by adding a macro parameter to the utility wrapper
and handling the cases by hand for each of the math functions. This
yields proper saturation arithmetic.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>