Add automatic unit tests for all 280+ functions
This was a major undertaking, but it is essential to have full test
coverage to not only catch regressions, but also validate the
correctness of the implementation.

There needed to be multiple rewrites to obtain the test infrastructure
in test/util.c. One can imagine further improvements, e.g. a randomised
approach for the large number types.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>