Fix two bugs in the codec
The automatic testing really proved useful in this regard, as it
uncovered significant arithmetic errors, e.g. when adding 1e-9 to
1.0, which yielded roughly 1.3 instead of 1.0.

After debugging it turned out to be two issues: Shifting an integer
by more than it's bit length is undefined behaviour, and gcc is almost
malicious by yielding complete garbage in that case. Thus we take
special care to check the shift amount in the mantissa bit preparation.

Additionally, a very subtle problem is a potential overflow to 1.0 of
m, which I totally overlooked. We handle this by manually checking for
it and accounting for it in c rather than m.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>