Make all operations exact up to 1 ULP
The use of a 'maximum relative error' for each type in the tests was a
bit of a hack and overshadowed the true necessity of coming within 1
ULP of a mostly exact reference result casted to the respective takum
type directly.

This has been reached now for both logarithmic and linear takums by
fixing many small things here and there. The automatic tests really made
this possible; despite the work of setting up the test environment they
really are a gift that keeps on giving.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>