Use base change for lb_1_plus() and lg_1_plus()
The original approach of just dumping more precision on the problem
was stupid, as changing the base of the logarithms from e to 2 and 10
respectively is as simple as multiplying with lb(e) and lg(e)
respectively, which we can store as constants.

This way we can use the stable log1pl implementation offered in math.h.

The unit tests showed that this make a big difference in many cases.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>