Use random sampling for non-exhaustive test cases
While our tests can fully cover 8- and (unary) 16-bit takum functions,
it is just not feasible for larger number number types. Previously we
would use a stride-length parameter to skip a specified number of
takums, however, this leads to repetitive bit patterns that might
obfuscate bugs in non-repetitive patterns.

The one thing that kept me from using random samples was that when the
number of test cases gets closer to the complete number of test cases
it gets more and more likely that we hit a case twice.

After a bit of research I stumbled upon linear-feedback shift registers
(LFSRs), which are a form of PRNG. Depending upon the choice of
'feedback polynomial' it is possible to determine the cycle length
of the generator exactly, in our case to 2^8-1, 2^16-1, 2^32-1 and
2^64-1. Only zero is not possible to 'hit', which is why it's the
starting value.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>