Uninline inline-defined functions in takum.h
While this provides the greatest flexibility to the C compiler, it
just grates on anyone intending to interface with the library from
other programming languages using FFIs. Given the inline functions
are not exported as library symbols, they are not available and force
the user to write their own implementations, which is not justified
despite some possible inlining-performance-optimisations.

This restructuring also helps cleaning up takum.h from internal
data structure definitions and improves the code readability, as
the user can just navigate to the respective files in src/ to
study a specific function implementation.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>