Refactor rounding code
Both src/codec.c and src/conversion.c had some kind of rounding
logic, which lead to duplicate code. As this is one of the more
critical arithmetic aspects it makes more sense to move this to one
very well-documented function in src/util.c that can handle both
use cases.

Additionally, instead of returning information on a carry that is
added separately, we just carry within the rounding and add the whole
result.

This drops some lines and shortens the pretty voluminous src/codec.c
significantly.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>